feat: add Jest configuration and dependencies to project
Add Jest configuration file and dependencies to the project. The configuration file sets up Jest to use the jsdom environment and imports the Next.js configuration. The package.json file has been updated to include the Jest dependency and scripts to run tests and watch tests. The @testing-library/jest-dom and @testing-library/react dependencies have also been added to the devDependencies.